@@ -17,7 +17,11 @@
   },
   "homepage": "https://github.com/Togira123/reactions-v2#readme",
   "dependencies": {
-    "discord.js": "^12.5.1",
-    "eslint": "^7.20.0"
+    "discord.js": "^13.6.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.10.0",
+    "ts-node": "^10.6.0",
+    "typescript": "^4.6.2"
   }
 }
